Rename : 사용페이지에 맞는 컴포넌트 트리정리
