+++ src/components/csr/Carousel.js
