+++ src/components/product/Carousel.module.scss
