+++ src/pages/product/bodyful/S04.js
@@ -5,8 +5,8 @@ import styles from './S04.module.scss';
 import product_ca from '../../../assets/images/product/bodyful/s04/product_ca.png';
 import product_ma from '../../../assets/images/product/bodyful/s04/product_ma.png';
 
-import Download from '../../../components/shared/Download';
-import Table from '../../../components/shared/Table';
+import Download from '../../../components/product/Download';
+import Table from '../../../components/product/Table';
 
 const cx = classNames.bind(styles);
